Split ast.cpp and main.cpp into several smaller files
