+++ src/ast.hpp
@@ -142,96 +142,5 @@ class Visitor {
   virtual ast_ptr Visit(Vec* node) = 0;
 };
 
-class Interp : public Visitor {
- public:
-  Interp();
-  Interp(const Example& world);
-  ast_ptr Visit(AST* node);
-  ast_ptr Visit(Var* node);
-  ast_ptr Visit(Param* node);
-  ast_ptr Visit(Num* node);
-  ast_ptr Visit(UnOp* node);
-  ast_ptr Visit(BinOp* node);
-  ast_ptr Visit(Vec* node);
-
- private:
-  Example world_;
-};
-
-class Print : public Visitor {
- public:
-  ast_ptr Visit(AST* node);
-  ast_ptr Visit(Num* node);
-  ast_ptr Visit(Var* node);
-  ast_ptr Visit(Param* node);
-  ast_ptr Visit(UnOp* node);
-  ast_ptr Visit(BinOp* node);
-  ast_ptr Visit(Vec* node);
-  void Display();
-
- private:
-  std::string program_ = "";
-  int depth_ = 0;
-};
-
-class ProblemGen : public Visitor {
- public:
-  ProblemGen(std::vector<Example>& example);
-  ast_ptr Visit(AST* node);
-  ast_ptr Visit(BinOp* node);
-  ast_ptr Visit(Num* node);
-  ast_ptr Visit(Param* node);
-  ast_ptr Visit(UnOp* node);
-  ast_ptr Visit(Var* node);
-  ast_ptr Visit(Vec* node);
-  std::string Get();
-
- private:
-  std::vector<std::string> assertions_;
-  std::vector<Example> examples_;
-  std::unordered_set<std::string> parameters_;
-};
-
-struct FunctionSig {
-  std::vector<Type> input_types_;
-  std::vector<Dimension> input_dims_;
-  Type output_type_;
-  Dimension output_dim_;
-};
-
-struct FunctionEntry {
-  std::string op_;
-  std::vector<Type> input_types_;
-  std::vector<Dimension> input_dims_;
-  Type output_type_;
-  Dimension output_dim_;
-};
-
-std::vector<ast_ptr> GetLegalOps(ast_ptr node, std::vector<ast_ptr> input,
-                                 const std::vector<FunctionEntry>& library);
-
-std::vector<ast_ptr> Enumerate(const std::vector<ast_ptr>& roots,
-                               const std::vector<ast_ptr>& inputs,
-                               const std::vector<FunctionEntry>& library);
-
-std::vector<ast_ptr> RecEnumerate(const std::vector<ast_ptr>& roots,
-                                  const std::vector<ast_ptr>& inputs,
-                                  const std::vector<Example>& examples,
-                                  const std::vector<FunctionEntry>& library,
-                                  const int depth,
-                                  std::vector<std::vector<float>>* signatures);
-
-template <typename T>
-bool IndexInVector(const std::vector<T>& vec, const T& element, int* index) {
-  // Find given element in vector
-  auto it = std::find(vec.begin(), vec.end(), element);
-  if (it != vec.end()) {
-    *index = distance(vec.begin(), it);
-    return true;
-  }
-  *index = -1;
-  return false;
-}
-
 }  // namespace AST
 #endif  // SRC_AST_HPP
